moving code to DeviceManager and setup for testing
